Fix garbled transaction ID in protos
Prior to this change, the default protobuf stringer would use the `Bytes`
method on `UUID` to get its hand on something it could (not actually) parse.

The real fix is to make the stringer smarter, added a TODO do that effect.

This change is a big improvement for the `cli/debug` comamnds.